+++ src/power_grid_model_io/config/excel/vision_en.yaml
@@ -424,8 +424,8 @@ grid:
       extra:
         - ID
         - Name
-      exclude_filter:
-        - power_grid_model_io.functions.exclude_filters.exclude_all_columns_empty_or_zero:
+      filters:
+        - power_grid_model_io.functions.filters.exclude_all_columns_empty_or_zero:
             cols: 
               - Load.P
               - Load.Q
@@ -449,8 +449,8 @@ grid:
         extra:
           - ID
           - Name
-        exclude_filter:
-          - power_grid_model_io.functions.exclude_filters.exclude_all_columns_empty_or_zero:
+        filters:
+          - power_grid_model_io.functions.filters.exclude_all_columns_empty_or_zero:
              cols: 
               - Generation.P
               - Generation.Q
@@ -476,8 +476,8 @@ grid:
         extra:
           - ID
           - Name
-        exclude_filter:
-          - power_grid_model_io.functions.exclude_filters.exclude_all_columns_empty_or_zero:
+        filters:
+          - power_grid_model_io.functions.filters.exclude_all_columns_empty_or_zero:
              cols: 
               - PV.Pnom
   Sources:
